+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [2.4.1] - 2026-05-21
+
+### Changed
+
+- MacGuffin drop embeds now put the drop type in the header and make the item name more prominent.
+- Updated the banana costume MacGuffin flavor text.
+
 ## [2.4.0] - 2026-05-21
 
 ### Added